+++ tests/fakes/NomadicPage.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Thinktomorrow\Chief\Tests\Fakes;
+
+use Thinktomorrow\Chief\Pages\Page;
+
+class NomadicPage extends Page
+{
+    protected static $managedModelKey = "nomadic_page";
+}